feat: Add functionality to delete videos
The code changes in `deleta-video.php` implement the functionality to delete videos from the database. When a video is deleted, the corresponding record is removed from the `videos` table. The deletion is triggered by the `id` parameter passed in the URL. If the deletion is successful, the user is redirected to the index page with a success message. Otherwise, they are redirected with a failure message.